backend/requirements.txt
asgiref==3.5.1
dj-rest-auth==2.1.3
Django==3.2
django-allauth==0.47.0
django-cors-headers==3.10.1
django-filter==21.1
django-rest-auth==0.9.5
django-social-share==2.2.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
oauthlib==3.1.1
PyJWT==2.3.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests-oauthlib==1.3.0
six==1.16.0